lib/requirements.txt
streamlit
spacy
networkx
pyvis
neo4j
sentence-transformers
faiss-cpu
numpy<2.0